Clean up commented code
